fixed bug with handling sys exec hash error code

+++ crates/core/src/evm/call.rs
@@ -30,7 +30,7 @@ pub fn _evm_call(input: EvmCallMethodInput) -> Result<Bytes, ExitCode> {
         callee_account.load_source_bytecode(),
     )))
     .unwrap();
-    let gas_limit = input.gas_limit as u64;
+    let gas_limit = input.gas_limit;
     // initiate contract instance and pass it to interpreter for and EVM transition
     let contract = Contract {
         input: input.input,
